fix: validate skill/interest existence and split opportunity-requirements service
